Splitted the MailerPlugin Currently I splitted the MailerPlugin so that everything that belongs together is inside a single file now. So Mailers => PlayMailer, SMTPConfiguration and Provider => SMTPConfiguration, Attachment => Attachment, MailerComponents => MailerComponents and Modules => MailerModule
